api/requirements.txt
annotated-types==0.7.0
asgiref==3.11.0
cffi==2.0.0
contextlib2==21.6.0
coverage==7.13.0
cryptography==46.0.3
Django==6.0
django-auditlog==3.4.1
django-cors-headers==4.6.0
django-mptt==0.16.0
django-ninja==1.5.1
django-ninja-extra==0.30.8
django-ninja-jwt==5.3.5
iniconfig==2.3.0
injector==0.23.0
model-bakery==1.20.5
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.10
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.2
pytest-cov==7.0.0
pytest-django==4.11.1
python-dateutil==2.9.0.post0
redis==5.2.1
six==1.17.0
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
